Refactor NIST task for upstream alignment and reduce code duplication
- Consolidate 3 observation files into single gear_insertion_observations.py
- Replace 4 custom obs functions with Isaac Lab built-ins (root_pos_w, root_quat_w)
- Slim NistGearInsertionTask constructor (40 → 17 params) via GraspConfig dataclass
- Hardcode reward weights in configclass instead of passing through constructor
- Delete bespoke play_rl_games.py; use generic policy_runner.py for evaluation
- Genericise RlGamesActionPolicy (remove NIST-specific defaults)
- Move RL-Games YAML config to isaaclab_arena_examples/policy/
- Clean up mdp/__init__.py re-exports
- Add merge readiness report and NIST vs Lift comparison doc
