--- isaaclab_arena/scripts/reinforcement_learning/play_rl_games.py
@@ -1,133 +1,0 @@
-# Copyright (c) 2025-2026, The Isaac Lab Arena Project Developers.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Play/evaluate an RL-Games checkpoint using the Arena environment builder.
-
-Launch Isaac Sim Simulator first.
-"""
-
-import math
-from pathlib import Path
-
-from isaaclab.app import AppLauncher
-
-from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
-from isaaclab_arena_environments.cli import add_example_environments_cli_args
-
-parser = get_isaaclab_arena_cli_parser()
-parser.add_argument("--video", action="store_true", default=False, help="Record videos during playback.")
-parser.add_argument("--video_length", type=int, default=200, help="Length of the recorded video (in steps).")
-parser.add_argument(
-    "--agent_cfg_path",
-    type=Path,
-    default=Path("isaaclab_arena/policy/rl_policy/nist_gear_insertion_osc_rl_games.yaml"),
-    help="Path to the RL-Games agent YAML configuration file.",
-)
-parser.add_argument("--checkpoint", type=str, required=True, help="Path to model checkpoint.")
-parser.add_argument("--sigma", type=str, default=None, help="Override the policy's standard deviation.")
-add_example_environments_cli_args(parser)
-args_cli = parser.parse_args()
-
-if args_cli.video:
-    args_cli.enable_cameras = True
-
-if getattr(args_cli, "enable_pinocchio", False):
-    import pinocchio  # noqa: F401
-
-app_launcher = AppLauncher(args_cli)
-simulation_app = app_launcher.app
-
-"""Rest everything follows."""
-
-import gymnasium as gym
-import os
-import yaml
-
-import omni.log
-from rl_games.common import env_configurations, vecenv
-from rl_games.common.algo_observer import IsaacAlgoObserver
-from rl_games.torch_runner import Runner
-
-from isaaclab.envs import DirectMARLEnv, multi_agent_to_single_agent
-from isaaclab.utils.assets import retrieve_file_path
-from isaaclab.utils.dict import print_dict
-
-from isaaclab_rl.rl_games import RlGamesGpuEnv, RlGamesVecEnvWrapper
-
-from isaaclab_arena_environments.cli import get_arena_builder_from_cli
-
-
-def main():
-    """Play with RL-Games agent."""
-    try:
-        arena_builder = get_arena_builder_from_cli(args_cli)
-        env_name, env_cfg = arena_builder.build_registered()
-    except Exception as e:
-        omni.log.error(f"Failed to parse environment configuration: {e}")
-        exit(1)
-
-    with open(args_cli.agent_cfg_path) as f:
-        agent_cfg = yaml.safe_load(f)
-
-    if args_cli.num_envs is not None:
-        env_cfg.scene.num_envs = args_cli.num_envs
-    if args_cli.device is not None:
-        env_cfg.sim.device = args_cli.device
-        agent_cfg["params"]["config"]["device"] = args_cli.device
-        agent_cfg["params"]["config"]["device_name"] = args_cli.device
-
-    resume_path = retrieve_file_path(args_cli.checkpoint)
-    agent_cfg["params"]["load_checkpoint"] = True
-    agent_cfg["params"]["load_path"] = resume_path
-    print(f"[INFO]: Loading model checkpoint from: {resume_path}")
-
-    env_cfg.seed = agent_cfg["params"]["seed"]
-
-    log_dir = os.path.dirname(resume_path)
-    env_cfg.log_dir = log_dir
-
-    env = gym.make(env_name, cfg=env_cfg, render_mode="rgb_array" if args_cli.video else None)
-
-    if isinstance(env.unwrapped, DirectMARLEnv):
-        env = multi_agent_to_single_agent(env)
-
-    if args_cli.video:
-        video_kwargs = {
-            "video_folder": os.path.join(log_dir, "videos", "play"),
-            "step_trigger": lambda step: step == 0,
-            "video_length": args_cli.video_length,
-            "disable_logger": True,
-        }
-        print("[INFO] Recording videos during playback.")
-        print_dict(video_kwargs, nesting=4)
-        env = gym.wrappers.RecordVideo(env, **video_kwargs)
-
-    rl_device = agent_cfg["params"]["config"]["device"]
-    clip_obs = agent_cfg["params"]["env"].get("clip_observations", math.inf)
-    clip_actions = agent_cfg["params"]["env"].get("clip_actions", math.inf)
-    obs_groups = agent_cfg["params"]["env"].get("obs_groups")
-    concate_obs_groups = agent_cfg["params"]["env"].get("concate_obs_groups", True)
-
-    env = RlGamesVecEnvWrapper(env, rl_device, clip_obs, clip_actions, obs_groups, concate_obs_groups)
-
-    vecenv.register(
-        "IsaacRlgWrapper",
-        lambda config_name, num_actors, **kwargs: RlGamesGpuEnv(config_name, num_actors, **kwargs),
-    )
-    env_configurations.register("rlgpu", {"vecenv_type": "IsaacRlgWrapper", "env_creator": lambda **kwargs: env})
-
-    agent_cfg["params"]["config"]["num_actors"] = env.unwrapped.num_envs
-
-    runner = Runner(IsaacAlgoObserver())
-    runner.load(agent_cfg)
-    runner.reset()
-
-    play_sigma = float(args_cli.sigma) if args_cli.sigma is not None else None
-    runner.run({"train": False, "play": True, "sigma": play_sigma, "checkpoint": resume_path})
-
-    env.close()
-
-
-if __name__ == "__main__":
-    main()
-    simulation_app.close()
