+++ isaaclab_arena_environments/nist_assembled_gearmesh_osc_environment.py
@@ -18,26 +18,30 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
     name: str = "nist_assembled_gear_mesh_osc"
 
     def get_env(self, args_cli: argparse.Namespace):
-        import isaaclab.envs.mdp as mdp_isaac_lab
         import isaaclab.sim as sim_utils
         from isaaclab.controllers import OperationalSpaceControllerCfg
         from isaaclab.envs.mdp.actions import BinaryJointPositionActionCfg
-        from isaaclab.managers import EventTermCfg, ObservationTermCfg as ObsTerm, SceneEntityCfg
+        from isaaclab.managers import ObservationTermCfg as ObsTerm
         from isaaclab.sensors.frame_transformer.frame_transformer_cfg import FrameTransformerCfg, OffsetCfg
 
         from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
+        from isaaclab_arena.reinforcement_learning.frameworks import RLFramework
         from isaaclab_arena.scene.scene import Scene
-        from isaaclab_arena.tasks.nist_gear_insertion_task import NistGearInsertionTask
+        from isaaclab_arena.tasks.nist_gear_insertion_task import GraspConfig, NistGearInsertionTask
         from isaaclab_arena.utils.pose import Pose
         import isaaclab_arena_environments.mdp as mdp
+        from isaaclab_arena.tasks.observations.gear_insertion_observations import (
+            NistGearInsertionPolicyObservations,
+        )
+        from isaaclab_arena_environments.mdp.nist_gear_insertion_osc_action import (
+            NistGearInsertionOscActionCfg,
+        )
 
         peg_tip_offset = (0.02025, 0.0, 0.025)
         peg_base_offset = (0.02025, 0.0, 0.0)
         success_z_fraction = 0.05
         xy_threshold = 0.0025
         episode_length_s = 15.0
-        pos_action_threshold = 0.02
-        rot_action_threshold = 0.097
 
         table = self.asset_registry.get_asset_by_name("table")()
         assembled_board = self.asset_registry.get_asset_by_name("nist_assembled_board")()
@@ -85,7 +89,7 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
                 0.04,
             ]
         )
-        embodiment.action_config.arm_action = mdp.NistGearInsertionOscActionCfg(
+        embodiment.action_config.arm_action = NistGearInsertionOscActionCfg(
             asset_name="robot",
             joint_names=["panda_joint[1-7]"],
             body_name="panda_fingertip_centered",
@@ -114,6 +118,9 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
             close_command_expr={"panda_finger_joint1": 0.0, "panda_finger_joint2": 0.0},
         )
 
+        # OSC insertion requires a specialized 24-D policy observation (peg-relative
+        # EE pose, force feedback, prev actions) instead of the default embodiment obs.
+        # The task_obs group adds privileged state for the critic separately.
         embodiment.observation_config.policy.actions = None
         embodiment.observation_config.policy.gripper_pos = None
         embodiment.observation_config.policy.eef_pos = None
@@ -121,7 +128,7 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
         embodiment.observation_config.policy.joint_pos = None
         embodiment.observation_config.policy.joint_vel = None
         embodiment.observation_config.policy.nist_gear_policy_obs = ObsTerm(
-            func=mdp.NistGearInsertionPolicyObservations,
+            func=NistGearInsertionPolicyObservations,
             params={
                 "robot_name": "robot",
                 "board_name": gears_and_base.name,
@@ -133,6 +140,8 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
                 "force_noise_level": 0.0,
             },
         )
+        # OSC torque control uses task-specific penalties (action magnitude, jerk,
+        # contact force) instead of the default joint-level regularizers.
         embodiment.reward_config.action_rate = None
         embodiment.reward_config.joint_vel = None
 
@@ -149,76 +158,7 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
         )
         scene = Scene(assets=[table, assembled_board, medium_gear, gears_and_base, light])
 
-        nist_randomization_events = {
-            "held_physics_material": EventTermCfg(
-                func=mdp_isaac_lab.randomize_rigid_body_material,
-                mode="startup",
-                params={
-                    "asset_cfg": SceneEntityCfg(medium_gear.name),
-                    "static_friction_range": (0.75, 0.75),
-                    "dynamic_friction_range": (0.75, 0.75),
-                    "restitution_range": (0.0, 0.0),
-                    "num_buckets": 1,
-                },
-            ),
-            "robot_physics_material": EventTermCfg(
-                func=mdp_isaac_lab.randomize_rigid_body_material,
-                mode="startup",
-                params={
-                    "asset_cfg": SceneEntityCfg("robot", body_names=".*"),
-                    "static_friction_range": (0.75, 0.75),
-                    "dynamic_friction_range": (0.75, 0.75),
-                    "restitution_range": (0.0, 0.0),
-                    "num_buckets": 1,
-                },
-            ),
-            "fixed_physics_material": EventTermCfg(
-                func=mdp_isaac_lab.randomize_rigid_body_material,
-                mode="startup",
-                params={
-                    "asset_cfg": SceneEntityCfg(gears_and_base.name),
-                    "static_friction_range": (0.25, 1.25),
-                    "dynamic_friction_range": (0.25, 0.25),
-                    "restitution_range": (0.0, 0.0),
-                    "num_buckets": 128,
-                },
-            ),
-            "held_object_mass": EventTermCfg(
-                func=mdp_isaac_lab.randomize_rigid_body_mass,
-                mode="reset",
-                params={
-                    "asset_cfg": SceneEntityCfg(medium_gear.name),
-                    "mass_distribution_params": (-0.005, 0.005),
-                    "operation": "add",
-                    "distribution": "uniform",
-                },
-            ),
-            "fixed_asset_pose": EventTermCfg(
-                func=mdp_isaac_lab.reset_root_state_uniform,
-                mode="reset",
-                params={
-                    "asset_cfg": SceneEntityCfg(gears_and_base.name),
-                    "pose_range": {
-                        "x": (0.0, 0.0),
-                        "y": (0.0, 0.0),
-                        "z": (0.0, 0.0),
-                        "roll": (0.0, 0.0),
-                        "pitch": (0.0, 0.0),
-                        "yaw": (0.0, 0.2617993877991494),
-                    },
-                    "velocity_range": {},
-                },
-            ),
-        }
-
-        task = NistGearInsertionTask(
-            assembled_board=assembled_board,
-            held_gear=medium_gear,
-            background_scene=table,
-            peg_offset_from_board=list(peg_base_offset),
-            success_z_fraction=success_z_fraction,
-            xy_threshold=xy_threshold,
-            start_in_gripper=True,
+        grasp_cfg = GraspConfig(
             num_arm_joints=7,
             hand_grasp_width=0.03,
             hand_close_width=0.0,
@@ -226,20 +166,23 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
             end_effector_body_name="panda_hand",
             grasp_rot_offset=[1.0, 0.0, 0.0, 0.0],
             grasp_offset=[0.02, 0.0, -0.128],
-            episode_length_s=episode_length_s,
-            enable_randomization=True,
             arm_joint_names="panda_joint[1-7]",
             finger_body_names=".*finger",
-            peg_offset_xy_noise=0.005,
-            gear_base_asset=gears_and_base,
+        )
+
+        task = NistGearInsertionTask(
+            assembled_board=assembled_board,
+            held_gear=medium_gear,
+            background_scene=table,
+            peg_offset_from_board=list(peg_base_offset),
             peg_offset_for_obs=list(peg_tip_offset),
-            include_insertion_regularizers=True,
-            pos_action_threshold=pos_action_threshold,
-            rot_action_threshold=rot_action_threshold,
-            engagement_xy_threshold=xy_threshold,
-            success_bonus_weight=1.0,
-            disable_drop_terminations=True,
-            extra_event_terms=nist_randomization_events,
+            gear_base_asset=gears_and_base,
+            success_z_fraction=success_z_fraction,
+            xy_threshold=xy_threshold,
+            episode_length_s=episode_length_s,
+            grasp_cfg=grasp_cfg,
+            enable_randomization=True,
+            rl_training_mode=args_cli.rl_training_mode,
         )
 
         return IsaacLabArenaEnvironment(
@@ -249,6 +192,8 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
             task=task,
             teleop_device=teleop_device,
             env_cfg_callback=mdp.assembly_env_cfg_callback,
+            rl_framework=RLFramework.RL_GAMES,
+            rl_policy_cfg="isaaclab_arena_examples.policy:nist_gear_insertion_osc_rl_games.yaml",
         )
 
     @staticmethod
@@ -256,4 +201,9 @@ class NISTAssembledGearMeshOSCEnvironment(ExampleEnvironmentBase):
         parser.add_argument("--embodiment", type=str, default="franka_ik", help="Robot embodiment")
         parser.add_argument(
             "--teleop_device", type=str, default=None, help="Teleoperation device (e.g., keyboard, spacemouse)"
+        )
+        parser.add_argument(
+            "--rl_training_mode",
+            action="store_true",
+            help="Disable success termination (use when training with RL-Games).",
         )
